requirements.txt
beautifulsoup4
click
pandas
requests
scipy
numpy
python-dateutil
matplotlib
SciencePlots